requirements.txt
python-telegram-bot==20.8
rembg==2.0.57
Pillow==10.4.0
